ajout de MancheService et PartieService (initMotATrouver placé dans MancheService)
