+++ src/main/java/fr/esgi/service/PartieService.java
@@ -1,0 +1,10 @@
+package fr.esgi.service;
+
+import fr.esgi.business.Partie;
+
+public interface PartieService {
+
+    Partie getPartie();
+    void setPartie(Partie partie);
+    void lancerNouvellePartie();
+}
